chore: minor design changes
