+++ shared/components/molecules/FeaturePages/ComparisonTable/index.ts
@@ -1,0 +1,2 @@
+export { default } from './ComparisonTable.view'
+export * from './ComparisonTable.types'
